+++ BraintreeCore/src/main/java/com/braintreepayments/api/AnalyticsEvent.kt
@@ -1,10 +1,9 @@
 package com.braintreepayments.api
 
-internal data class AnalyticsEvent constructor(
+internal data class AnalyticsEvent(
     val name: String,
     val payPalContextId: String? = null,
     val linkType: String? = null,
-    val venmoInstalled: Boolean = false,
     val isVaultRequest: Boolean = false,
     val startTime: Long? = null,
     val endTime: Long? = null,