+++ BraintreeCore/src/main/java/com/braintreepayments/api/BraintreeClient.kt
@@ -53,7 +53,6 @@ open class BraintreeClient @VisibleForTesting internal constructor(
 
     private val crashReporter: CrashReporter
     private var launchesBrowserSwitchAsNewTask: Boolean = false
-    private val deviceInspector: DeviceInspector
 
     // NOTE: this constructor is used to make dependency injection easy
     internal constructor(params: BraintreeClientParams) : this(
@@ -151,7 +150,6 @@ open class BraintreeClient @VisibleForTesting internal constructor(
         // statistics access via the sdk console
         crashReporter = CrashReporter(this)
         crashReporter.start()
-        deviceInspector = DeviceInspector()
     }
 
     /**
@@ -196,12 +194,10 @@ open class BraintreeClient @VisibleForTesting internal constructor(
         getAuthorization { authorization, _ ->
             if (authorization != null) {
                 getConfiguration { configuration, _ ->
-                    val isVenmoInstalled = deviceInspector.isVenmoInstalled(applicationContext)
                     val event = AnalyticsEvent(
                         eventName,
                         params.payPalContextId,
                         params.linkType,
-                        isVenmoInstalled,
                         params.isVaultRequest,
                         params.startTime,
                         params.endTime,
@@ -319,12 +315,14 @@ open class BraintreeClient @VisibleForTesting internal constructor(
                         ) { response, httpError ->
                             response?.let {
                                 try {
-                                    json?.optString(GraphQLConstants.Keys.OPERATION_NAME)
+                                    json?.optString(GraphQLConstants.Keys.QUERY)
                                         ?.let { query ->
+                                            val queryDiscardHolder = query.replace(Regex("^[^\\(]*"), "")
+                                            val finalQuery = query.replace(queryDiscardHolder, "")
                                             val params = AnalyticsEventParams(
                                                 startTime = it.timing.startTime,
                                                 endTime = it.timing.endTime,
-                                                endpoint = query
+                                                endpoint = finalQuery
                                             )
                                             sendAnalyticsEvent(
                                                 CoreAnalytics.apiRequestLatency,
@@ -516,7 +514,11 @@ open class BraintreeClient @VisibleForTesting internal constructor(
     }
 
     private fun sendAnalyticsTimingEvent(endpoint: String, timing: HttpResponseTiming) {
-        val cleanedPath = endpoint.replace(Regex("/merchants/([A-Za-z0-9]+)/client_api"), "")
+        var cleanedPath = endpoint.replace(Regex("/merchants/([A-Za-z0-9]+)/client_api"), "")
+        cleanedPath = cleanedPath.replace(
+            Regex("payment_methods/.*/three_d_secure"), "payment_methods/three_d_secure"
+        )
+
         sendAnalyticsEvent(
             CoreAnalytics.apiRequestLatency,
             AnalyticsEventParams(